Version 1
Major functionality are completed in this version
There are 7 recipes added to the app
It can show recipe info and pair with image
What we need to implement next should be the filter function.
Current revision is very useless

+++ RecipeManager.py
@@ -95,4 +95,4 @@ class RecipeManager:
             for row in reader:
                 name, ingredients, instructions, rating, image_paths = row
                 images = [Image(path.strip()) for path in image_paths.split(',')]
-                self.addRecipe(Recipe(name, ingredients, instructions,int(rating), images))
+                self.addRecipe(Recipe(name, ingredients, instructions,float(rating), images))
